Handle IO for file/STDIN/STDOUT elegantly
