+++ src/lib/io.rs
@@ -1,18 +1,8 @@
 //! A module for commonly used IO functions.
 use std::ffi::OsStr;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 use crate::path::GZIP_EXTENSION;
-
-/// A path to this system's standard input file descriptor.
-pub fn stdin() -> PathBuf {
-    PathBuf::from("/dev/stdin")
-}
-
-/// A path to this system's standard output file descriptor.
-pub fn stdout() -> PathBuf {
-    PathBuf::from("/dev/stdout")
-}
 
 /// Test if a path has a gzip extension or not.
 pub fn has_gzip_ext<P>(path: &P) -> bool
@@ -25,17 +15,9 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::path::PathBuf;
+
     use super::*;
-
-    #[test]
-    fn test_stdin() {
-        assert_eq!(stdin(), PathBuf::from("/dev/stdin"));
-    }
-
-    #[test]
-    fn test_stdout() {
-        assert_eq!(stdout(), PathBuf::from("/dev/stdout"));
-    }
 
     #[test]
     fn test_has_gzip_ext() {
