+++ src/main.rs
@@ -1,15 +1,18 @@
 //! Convert variants from VarDict/VarDictJava into VCF format, fast.
 //! Only output from tumor-only calling is currently supported.
+use std::fs::File;
+use std::io::{BufReader, Read};
 use std::path::PathBuf;
 use std::process;
 
+use anyhow::{Error, Result};
 use env_logger::Env;
+use log::*;
 use structopt::StructOpt;
 
-use vartovcflib::io;
 use vartovcflib::vartovcf;
 
-#[derive(Clone, StructOpt, Debug)]
+#[derive(Clone, Debug, StructOpt)]
 #[structopt(rename_all = "kebab-case", about)]
 struct Opt {
     /// The sample name
@@ -30,14 +33,29 @@ struct Opt {
 }
 
 /// Main binary entrypoint.
-fn main() {
+#[cfg(not(tarpaulin_include))]
+fn main() -> Result<(), Error> {
     let env = Env::default().default_filter_or(vartovcflib::DEFAULT_LOG_LEVEL);
     let opt = Opt::from_args();
-    let input = opt.input.unwrap_or_else(io::stdin);
-    let output = opt.output.unwrap_or_else(io::stdout);
 
     env_logger::Builder::from_env(env).init();
-    match vartovcf::run(&input, &output, &opt.reference, &opt.sample) {
+
+    let input: Box<dyn Read> = match &opt.input {
+        Some(path) if path.to_str().unwrap() != "-" => {
+            info!("Input file: {:?}", path);
+            Box::new(BufReader::new(File::open(path)?))
+        }
+        _ => {
+            info!("Input stream: STDIN");
+            Box::new(std::io::stdin())
+        }
+    };
+    match &opt.output {
+        Some(output) => info!("Output file: {:?}", output),
+        None => info!("Output stream: STDOUT"),
+    }
+
+    match vartovcf(input, opt.output, &opt.reference, &opt.sample) {
         Ok(exit_code) => process::exit(exit_code),
         Err(except) => panic!("{}", except),
     }
